refactor: move Websocket and Discord services from app.py to services folder
